moved shared amounts types from epic to shared type definition file
