+++ packages/modules/src/hooks/useChoiceCards.ts
@@ -1,7 +1,10 @@
-import { ContributionFrequency, SelectedAmountsVariant } from '@sdc/shared/src/types/abTests/epic';
 import { useState, useEffect } from 'react';
 import { BannerTextContent } from '../modules/banners/common/types';
 import { getLocalCurrencySymbol } from '@sdc/shared/dist/lib';
+import {
+    ContributionFrequency,
+    SelectedAmountsVariant,
+} from '@sdc/shared/src/types/abTests/shared';
 
 export interface ChoiceCardSelection {
     frequency: ContributionFrequency;
